solver base prepare_eval_grad change function variable name from path to cwd to avoid confusion with the workflow path that is also passed in LBFGS changed default of LBFGSMAX to inf from null readers and writers now have working ASCII components for preprocess slurm lg ENVIRON is now set to empty str and not NoneType

+++ seisflows/postprocess/base.py
@@ -44,7 +44,7 @@ class Base:
                docstr="Gaussian half-width for vertical smoothing in units "
                       "of meters")
 
-        sf.par("TASKTIME_SMOOTH", required=False, default=1., par_type=float,
+        sf.par("TASKTIME_SMOOTH", required=False, default=1, par_type=int,
                docstr="Large radii smoothing may take longer than normal "
                       "tasks. Allocate additional smoothing task time "
                       "as a multiple of TASKTIME")
